Logic update for scanning
I believe you had the scan checking for a 20% chance of success, which
is the opposite of what is intended. I switched it so that it was random
from 0-5, and any number less than 4 would be a success. It also kind of
works now, it seems the logic for checking if someone goes to jail is
working.